:sparkles: api: separate sco and verification-code parcoursup seeds
