+++ api/datamart/seeds/cases/verification-code/verification-code-only.js
@@ -9,20 +9,21 @@ import {
   generateStatus,
   getCertificationDate,
   getFormattedBirthdate,
-  nationalStudentIdGenerator,
-} from './tools.js';
+  verificationCodeGenerator,
+} from '../tools.js';
 
-const generateCertifCourseId = certificationCourseIdGenerator({ startingFrom: 1100000 });
-const generateINE = nationalStudentIdGenerator({ ineSuffix: 'AA' });
+const generateCertifCourseId = certificationCourseIdGenerator({ startingFrom: 1000000 });
+const generateVerificationCode = verificationCodeGenerator({ startingFrom: 10000000 });
 
 /**
- * A student that has a V3 certification and that can be found by INE but not by UAI
+ * A person that has a certification not linked to an INE or UAI
  */
 export default function () {
   const studentBase = {
     certification_courses_id: generateCertifCourseId(),
-    organization_uai: null, // We do not want it to be accessible by UAI
-    national_student_id: generateINE(),
+    certification_code_verification: generateVerificationCode(),
+    organization_uai: null,
+    national_student_id: null,
     last_name: faker.person.lastName(),
     first_name: generateFirstName(),
     birthdate: getFormattedBirthdate(),
