+++ api/package.json
@@ -138,6 +138,8 @@
     "datamart:rollback:latest": "knex --knexfile datamart/knexfile.js migrate:down",
     "datamart:prepare": "run-s datamart:delete datamart:create datamart:migrate",
     "datamart:seed": "NODE_OPTIONS=\"--max-old-space-size=380\" knex --knexfile datamart/knexfile.js seed:run",
+    "datamart:seed:sco": "NODE_OPTIONS=\"--max-old-space-size=380\" knex --knexfile datamart/knexfile.js seed:run --specific=populate-sco-datamart.js",
+    "datamart:seed:verification-code": "NODE_OPTIONS=\"--max-old-space-size=380\" knex --knexfile datamart/knexfile.js seed:run --specific=populate-verification-code-datamart.js",
     "datamart:reset": "npm run datamart:prepare && npm run datamart:seed",
     "datawarehouse:create": "node scripts/database/create-database --name datawarehouse",
     "datawarehouse:delete": "node scripts/database/drop-database --name datawarehouse",
